test: remove disabled features code duplicated
