+++ bigbluebutton-tests/playwright/parameters/customparameters.js
@@ -230,201 +230,6 @@ class CustomParameters extends MultiUsers {
     expect(await this.modPage.getLocator(e.selectCameraQualityId).inputValue()).toBe('low');
     await this.modPage.waitAndClick(e.startSharingWebcam);
   }
-
-  async breakoutRooms() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.wasRemoved(e.createBreakoutRooms);
-  }
-
-  async speechRecognition() {
-    await this.modPage.waitForSelector(e.audioModal, ELEMENT_WAIT_LONGER_TIME);
-    await this.modPage.wasRemoved(e.speechRecognition);
-  }
-
-  async captions() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.wasRemoved(e.writeClosedCaptions);
-  }
-
-  async chat() {
-    await this.modPage.wasRemoved(e.publicChat);
-  }
-
-  async externalVideos() {
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.wasRemoved(e.shareExternalVideoBtn);
-  }
-
-  async layouts() {
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.wasRemoved(e.layoutModal);
-  }
-
-  async learningDashboard() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.wasRemoved(e.learningDashboard);
-  }
-
-  async polls() {
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.wasRemoved(e.polling);
-  }
-
-  async screenshare() {
-    await this.modPage.wasRemoved(e.startScreenSharing);
-  }
-
-  async sharedNotes() {
-    await this.modPage.wasRemoved(e.sharedNotes);
-  }
-
-  async virtualBackgrounds() {
-    await this.modPage.waitAndClick(e.joinVideo);
-    await this.modPage.wasRemoved(e.virtualBackgrounds);
-  }
-
-  async downloadPresentationWithAnnotations() {
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.waitAndClick(e.managePresentations);
-    await this.modPage.waitAndClick(e.presentationOptionsDownloadBtn);
-    await this.modPage.wasRemoved(e.sendPresentationInCurrentStateBtn);
-  }
-
-  async importPresentationWithAnnotationsFromBreakoutRooms() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.waitAndClick(e.createBreakoutRooms);
-    await this.modPage.wasRemoved(e.captureBreakoutWhiteboard);
-  }
-
-  async importSharedNotesFromBreakoutRooms() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.waitAndClick(e.createBreakoutRooms);
-    await this.modPage.wasRemoved(e.captureBreakoutSharedNotes);
-  }
-
-  async presentation() {
-    await this.modPage.wasRemoved(e.whiteboard);
-    await this.modPage.wasRemoved(e.minimizePresentation);
-    await this.modPage.wasRemoved(e.restorePresentation);
-  }
-
-  async customVirtualBackground() {
-    await this.modPage.waitAndClick(e.joinVideo);
-    await this.modPage.waitForSelector(e.webcamSettingsModal);
-    await this.modPage.wasRemoved(e.inputBackgroundButton);
-  }
-
-  async slideSnapshot() {
-    await this.modPage.waitForSelector(e.whiteboard);
-    await this.modPage.waitAndClick(e.whiteboardOptionsButton);
-    await this.modPage.hasElement(e.presentationFullscreen);
-    await this.modPage.wasRemoved(e.presentationSnapshot);
-  }
-
-  async cameraAsContent() {
-    await this.modPage.waitForSelector(e.whiteboard);
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.hasElement(e.managePresentations);
-    await this.modPage.wasRemoved(e.shareCameraAsContent);
-  }
-
-  // Disabled Features Exclude
-  async breakoutRoomsExclude() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.hasElement(e.createBreakoutRooms);
-  }
-
-  async speechRecognitionExclude() {
-    const { speechRecognitionEnabled } = getSettings();
-    test.fail(!speechRecognitionEnabled, 'Live Transcription is disabled');
-    await this.modPage.waitForSelector(e.audioModal, ELEMENT_WAIT_LONGER_TIME);
-    await this.modPage.hasElement(e.speechRecognition);
-  }
-
-  async captionsExclude() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.hasElement(e.writeClosedCaptions);
-  }
-
-  async chatExclude() {
-    await this.modPage.hasElement(e.publicChat);
-  }
-
-  async externalVideosExclude() {
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.hasElement(e.shareExternalVideoBtn);
-  }
-
-  async layoutsExclude() {
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.hasElement(e.layoutModal);
-  }
-
-  async learningDashboardExclude() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.hasElement(e.learningDashboard);
-  }
-
-  async pollsExclude() {
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.hasElement(e.polling);
-  }
-
-  async screenshareExclude() {
-    await this.modPage.hasElement(e.startScreenSharing);
-  }
-
-  async sharedNotesExclude() {
-    await this.modPage.hasElement(e.sharedNotes);
-  }
-
-  async virtualBackgroundsExclude() {
-    await this.modPage.waitAndClick(e.joinVideo);
-    await this.modPage.hasElement(e.virtualBackgrounds);
-  }
-
-  async downloadPresentationWithAnnotationsExclude() {
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.waitAndClick(e.managePresentations);
-    await this.modPage.waitAndClick(e.presentationOptionsDownloadBtn);
-    await this.modPage.hasElement(e.sendPresentationInCurrentStateBtn);
-  }
-
-  async importPresentationWithAnnotationsFromBreakoutRoomsExclude() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.waitAndClick(e.createBreakoutRooms);
-    await this.modPage.hasElement(e.captureBreakoutWhiteboard);
-  }
-
-  async importSharedNotesFromBreakoutRoomsExclude() {
-    await this.modPage.waitAndClick(e.manageUsers);
-    await this.modPage.waitAndClick(e.createBreakoutRooms);
-    await this.modPage.hasElement(e.captureBreakoutSharedNotes);
-  }
-
-  async presentationExclude() {
-    await this.modPage.hasElement(e.whiteboard);
-    await this.modPage.waitAndClick(e.minimizePresentation);
-    await this.modPage.hasElement(e.restorePresentation);
-  }
-
-  async customVirtualBackgroundExclude() {
-    await this.modPage.waitAndClick (e.joinVideo);
-    await this.modPage.waitForSelector(e.webcamSettingsModal);
-    await this.modPage.hasElement(e.inputBackgroundButton);
-  }
-
-  async slideSnapshotExclude() {
-    await this.modPage.waitForSelector(e.whiteboard);
-    await this.modPage.waitAndClick(e.whiteboardOptionsButton);
-    await this.modPage.hasElement(e.presentationSnapshot);
-  }
-
-  async cameraAsContentExclude() {
-    await this.modPage.waitForSelector(e.whiteboard);
-    await this.modPage.waitAndClick(e.actions);
-    await this.modPage.hasElement(e.shareCameraAsContent);
-  }
 }
 
 exports.CustomParameters = CustomParameters;
