+++ bigbluebutton-tests/playwright/parameters/parameters.spec.js
@@ -312,25 +312,25 @@ test.describe.parallel('Create Parameters', () => {
     
     test.describe.serial(() => {
       test('Slide Snapshot', async ({ browser, context, page }) => {
-        const customParam = new CustomParameters(browser, context);
-        await customParam.initModPage(page, true, { customParameter: c.slideSnapshotDisabled });
-        await customParam.slideSnapshot();
+        const disabledFeatures = new DisabledFeatures(browser, context);
+        await disabledFeatures.initModPage(page, true, { customParameter: c.slideSnapshotDisabled });
+        await disabledFeatures.slideSnapshot();
       });
       test('Slide Snapshot (exclude)', async ({ browser, context, page }) => {
-        const customParam = new CustomParameters(browser, context);
-        await customParam.initModPage(page, true, { customParameter: c.slideSnapshotExclude });
-        await customParam.slideSnapshotExclude();
+        const disabledFeatures = new DisabledFeatures(browser, context);
+        await disabledFeatures.initModPage(page, true, { customParameter: c.slideSnapshotExclude });
+        await disabledFeatures.slideSnapshotExclude();
       });
     });
 
     test.describe.serial(() => {
       test('Camera As Content', async ({ browser, context, page }) => {
-        const customParam = new CustomParameters(browser, context);
+        const customParam = new DisabledFeatures(browser, context);
         await customParam.initModPage(page, true, { customParameter: c.cameraAsContent });
         await customParam.cameraAsContent();
       });
       test('Camera As Content (exclude)', async ({ browser, context, page }) => {
-        const customParam = new CustomParameters(browser, context);
+        const customParam = new DisabledFeatures(browser, context);
         await customParam.initModPage(page, true, { customParameter: c.cameraAsContentExclude });
         await customParam.cameraAsContentExclude();
       });
